// projects/html-builder-angular/src/lib/enums/element-alignment.enums.ts
export enum ElementAlignments {
  HORIZONTAL_LEFT = 'left',
  HORIZONTAL_CENTER = 'center',
  HORIZONTAL_RIGHT = 'right',
  HORIZONTAL_JUSTIFY = 'justify',
  VERTICAL_TOP = 'vertical_top',
  VERTICAL_CENTER = 'vertical_center',
  VERTICAL_BOTTOM = 'vertical_bottom',
  CENTER_CENTER = 'center_center',
}

export enum ElementAlignmentType {
  HORIZONTAL = 'horizontal',
  VERTICAL = 'vertical',
  CONTENT = 'content',
}